#76 refactor: to Unify the style of the components
