+++ src/components/My/MyPage/index.jsx
@@ -4,11 +4,14 @@ import Process from './Process'
 import MyPossession from './MyPossession'
 import MyNav from './MyNav'
 import Guest from './Guest'
-
-const MyPageSegment = () => {
+import Button from '../../common/Button'
+const index = () => {
   // api (useEffect async await)
   const token = localStorage.getItem('accessToken')
-
+  const logoutHandler = () => {
+    //로그아웃 api작성 예정
+    console.log('logout!!')
+  }
   return (
     <div>
       <Header />
@@ -21,9 +24,15 @@ const MyPageSegment = () => {
           <MyPossession />
         </>
       )}
+      <Button
+        onClick={logoutHandler}
+        classprop="text-sm mx-5 mt-9 mb-5 bg-primary text-white"
+      >
+        로그아웃
+      </Button>
       <MyNav />
     </div>
   )
 }
 
-export default MyPageSegment
+export default index
