+++ src/pages/My/MyPage.jsx
@@ -1,6 +1,5 @@
 import React from 'react'
-import Button from '../../components/common/Button'
-import MyPageSegment from '../../components/My/MyPageSegment'
+import { Outlet } from 'react-router-dom'
 const MyPage = () => {
   const logoutHandler = () => {
     //로그아웃 api작성 예정
@@ -9,13 +8,7 @@ const MyPage = () => {
   return (
     <>
       <div>
-        <MyPageSegment />
-        <Button
-          onClick={logoutHandler}
-          classprop="text-sm mx-5 mt-9 mb-5 bg-primary text-white"
-        >
-          로그아웃
-        </Button>
+        <Outlet />
       </div>
     </>
   )
