--- src/pages/My/MyReviewWrite.jsx
@@ -1,11 +1,0 @@
-import React from 'react'
-import Write from '../../components/My/MyReviewSegment/Write'
-const MyReviewWrite = () => {
-  return (
-    <>
-      <Write />
-    </>
-  )
-}
-
-export default MyReviewWrite
